# docs/web-api.openapi.yaml
openapi: 3.1.0
info:
  title: Omegon Web API
  version: 0.27.0
  description: >
    Browser-native control surface for the Omegon web dashboard ("omegon-web").
    Covers the browser-native Web UI API: `/api/web/*` compatibility routes,
    native first-party `/api/sessions*` session resources, daemon event ingress,
    capability discovery, launch context, semantic surface snapshots/streams,
    semantic action ingress, saved-session listing, and staged browser attachments.


    Derived from the axum handlers in `core/crates/omegon/src/web/` (api.rs,
    surfaces.rs, surface_stream.rs) and the UI-runtime outcome envelope in
    `core/crates/omegon/src/ui_runtime/envelope.rs`. This document is the
    source of truth for the wire shape — code implements the contract.
servers:
  - url: /
    description: Same-origin as the served dashboard (default 127.0.0.1 bind).
...
paths:
...
  /api/sessions/{session_id}/surfaces/stream:
    get:
      tags: [surfaces]
      operationId: streamNativeSessionSurfaces
      summary: Native session live surface stream (WebSocket upgrade).
      description: >
        Requires either a valid web-auth query token, or bearer auth plus optional trusted principal headers, and the `surface.stream` RBAC operation before the WebSocket upgrade is accepted.
      parameters:
        - name: session_id
          in: path
          required: true
          description: Session id. Only `default` is currently accepted.
          schema:
            type: string
        - name: token
          in: query
          required: false
          description: Web-auth query token.
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols — WebSocket established.
        "401":
          description: Missing or invalid web-auth token.
        "403":
          description: Caller role lacks `surface.stream`.
        "404":
          description: Unknown session_id.
      x-websocket-message:
        $ref: "#/components/schemas/SurfaceStreamMessage"
...
  /api/web/surfaces:
    get:
      tags: [surfaces]
      operationId: getWebSurfaces
      summary: Browser-native semantic surface snapshot.
      description: >
        Full point-in-time projection of every renderer-neutral surface for the
        live session.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Surface snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSurfacesSnapshot"
        "403":
          description: Caller role lacks `surface.read`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RbacErrorResponse"
  /api/web/surfaces/stream:
    get:
      tags: [surfaces]
      operationId: streamWebSurfaces
      summary: Live surface stream (WebSocket upgrade).
      description: >
        Upgrades the GET request to a WebSocket. Requires either a web-auth query token, or bearer auth plus optional trusted principal headers. On connect the server sends a
        `snapshot` message carrying the full `WebSurfacesSnapshot` as `payload`,
        then one message per agent event with a monotonically increasing
        `revision`. On consumer lag the server emits an `error` message advising
        a refetch of `/api/web/surfaces`. Messages are JSON text frames.
      parameters:
        - name: token
          in: query
          required: false
          description: >
            Web-auth query token. Required when the server has web auth enabled;
            an invalid or missing token yields `401` before the upgrade.
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols — WebSocket established.
        "401":
          description: Missing or invalid web-auth token.
        "403":
          description: Caller role lacks `surface.stream`.
      x-websocket-message:
        $ref: "#/components/schemas/SurfaceStreamMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    RbacErrorResponse:
      type: object
      required: [schema_version, error, reason, operation, capability, required_base, role, mode]
      properties:
        schema_version:
          type: integer
          const: 1
        error:
          type: string
          enum: [unauthorized, invalid_role, forbidden, misconfigured, policy_unavailable]
        reason:
          type: string
        operation:
          type: [string, "null"]
        capability:
          type: [string, "null"]
        required_base:
          type: [string, "null"]
        role:
          type: [string, "null"]
        mode:
          type: string
...
    WebSurfacesSnapshot:
      type: object
      required: [schema_version, session_id, revision, generated_at, surfaces]
      properties:
        schema_version:
          type: integer
          format: int32
          const: 1
        session_id: { type: string }
        revision:
          type: integer
          format: int64
          minimum: 0
        generated_at:
          type: string
          description: RFC 3339 timestamp.
        surfaces:
          $ref: "#/components/schemas/WebSurfaceBundle"
    WebSurfaceBundle:
      type: object
      description: >
        Renderer-neutral surfaces for one session. All surfaces are modeled in
        full from the DTOs in `core/crates/omegon/src/web/surfaces.rs`.
      required:
        - conversation
        - editor
        - command
        - command_menu
        - dashboard
        - footer
        - instruments
        - memory_status
        - operations
        - settings
      properties:
        conversation:
          $ref: "#/components/schemas/WebConversationSurface"
        editor:
          $ref: "#/components/schemas/WebEditorSurface"
        command:
          $ref: "#/components/schemas/WebCommandSurface"
        command_menu:
          $ref: "#/components/schemas/WebCommandMenuSurface"
        dashboard:
          $ref: "#/components/schemas/WebDashboardSurface"
        footer:
          $ref: "#/components/schemas/WebFooterSurface"
        instruments:
          $ref: "#/components/schemas/WebInstrumentsSurface"
        memory_status:
          $ref: "#/components/schemas/WebMemoryStatusSurface"
        operations:
          $ref: "#/components/schemas/WebOperationsSurface"
        settings:
          $ref: "#/components/schemas/WebSettingsSurface"
    WebEditorSurface:
      type: object
      required: [accepts_prompt, placeholder, queue_mode, supports_attachments]
      properties:
        accepts_prompt:
          type: boolean
          description: Whether the editor currently accepts a prompt submission.
        placeholder:
          type: string
        queue_mode:
          type: string
          description: >
            How prompts submitted while busy are handled (e.g. `until_ready`).
        supports_attachments:
          type: boolean
    WebCommandSurface:
      type: object
      properties:
        pending_prompt:
          type: [string, "null"]
          description: A prompt staged but not yet submitted, if any.
    WebCommandMenuSurface:
      type: object
      required: [available, open]
      properties:
        available:
          type: boolean
          description: Whether the slash-command menu can be opened.
        open:
          type: boolean
          description: Whether the menu is currently open.
    WebDashboardSurface:
      type: object
      required:
        - session
        - lifecycle_available
        - cleave_available
        - delegate_available
        - harness_available
      properties:
        session:
          $ref: "#/components/schemas/WebDashboardSessionSurface"
        lifecycle_available:
          type: boolean
        cleave_available:
          type: boolean
        delegate_available:
          type: boolean
        harness_available:
          type: boolean
    WebDashboardSessionSurface:
      type: object
      required: [turns, tool_calls, compactions]
      properties:
        turns:
          type: integer
          format: int32
          minimum: 0
        tool_calls:
          type: integer
          format: int32
          minimum: 0
        compactions:
          type: integer
          format: int32
          minimum: 0
    WebFooterSurface:
      type: object
      required: [busy]
      properties:
        busy:
          type: boolean
          description: Whether the agent is mid-turn.
    WebInstrumentsSurface:
      type: object
      properties:
        active_tool:
          type: [string, "null"]
          description: Name of the tool currently executing, if any.
    WebMemoryStatusSurface:
      type: object
      required: [active_facts, total_facts]
      properties:
        active_facts:
          type: integer
          minimum: 0
        total_facts:
          type: integer
          minimum: 0
    WebOperationsSurface:
      type: object
      required: [active_child_runtimes]
      properties:
        active_child_runtimes:
          type: integer
          minimum: 0
          description: Count of active delegate/child runtimes.
    WebSettingsSurface:
      type: object
      properties:
        auth_mode:
          type: [string, "null"]
        auth_source:
          type: [string, "null"]
    WebConversationSurface:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/WebConversationSegment"
    WebConversationSegment:
      type: object
      required: [index, role, complete, copyable, selectable]
      properties:
        index:
          type: integer
          minimum: 0
        role: { type: string }
        title:
          type: [string, "null"]
        summary:
          type: [string, "null"]
        body:
          type: [string, "null"]
        complete: { type: boolean }
        copyable: { type: boolean }
        selectable: { type: boolean }
    SurfaceStreamMessage:
      type: object
      description: One JSON text frame on the /api/web/surfaces/stream WebSocket.
      required: [schema_version, session_id, revision, type, payload]
      properties:
        schema_version:
          type: integer
          format: int32
          const: 1
        session_id: { type: string }
        revision:
          type: integer
          format: int64
        type:
          type: string
          description: >
            `snapshot` for the initial full projection, `error` on consumer lag,
            otherwise an agent-event-derived message type.
        surface:
          type: [string, "null"]
          description: Surface key the message pertains to, when scoped.
        payload:
          description: >
            For `snapshot`, a WebSurfacesSnapshot. For `error`, `{ "message": ... }`.
            Otherwise an agent-event-shaped object.
          oneOf:
            - $ref: "#/components/schemas/WebSurfacesSnapshot"
            - type: object
              additionalProperties: true
...